requirements.txt
Telethon==1.36.0
aiogram==3.10.0
SQLAlchemy==2.0.31
aiosqlite==0.20.0
python-dotenv==1.0.1
greenlet==3.0.3
PyYAML==6.0.1
